requirements.txt
fastapi==0.85.0
uvicorn==0.18.3
sqlalchemy==1.4.41
pydantic==1.10.2
pymysql==1.0.2
alembic==1.8.1
python-dotenv==0.21.0
python-jose==3.3.0
passlib==1.7.4
jinja2==3.1.2
cryptography==3.4.7
python-multipart==0.0.5
email-validator==1.1.2
bcrypt==3.2.0